booking_system_rest/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic[email]
python-dotenv
pytest
pytest-asyncio
pytest-cov
httpx
pytest-mock
PyJWT[crypto]
